+++ src/main/kotlin/net/purefunc/c2c/lottery/data/dto/response/OrderDtoRes.kt
@@ -13,7 +13,7 @@ data class OrderDtoRes(
 
     val multiple: Int,
 
-    var totalAmount: BigDecimal,
+    var totalAmount: Int,
 
     var winAmount: BigDecimal,
