+++ src/main/kotlin/net/purefunc/c2c/lottery/data/table/OrderDo.kt
@@ -19,7 +19,7 @@ data class OrderDo(
 
     var multiple: Int,
 
-    var totalAmount: BigDecimal,
+    var totalAmount: Int,
 
     var winAmount: BigDecimal,
